+++ core/design-system/src/main/java/com/msg/design_system/component/dialog/ApplicationRejectDialog.kt
@@ -14,12 +14,10 @@ import androidx.compose.foundation.shape.CornerSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
@@ -28,15 +26,15 @@ import com.msg.design_system.component.description.ContentDescriptionText
 import com.msg.design_system.theme.BitgoeulAndroidTheme
 
 @Composable
-fun BitgoeulApproveDialog(
+fun ApplicationReject(
     modifier: Modifier = Modifier,
     content: String,
     isVisible: Boolean,
     onQuit: () -> Unit,
 ) {
     if (isVisible) {
-        Dialog(onDismissRequest = { onQuit() }) {
-            BitgoeulAndroidTheme { colors, type ->
+        BitgoeulAndroidTheme { colors, type ->
+            Dialog(onDismissRequest = { onQuit() }) {
                 Box(
                     modifier = modifier
                         .background(
@@ -50,7 +48,7 @@ fun BitgoeulApproveDialog(
                     Column {
                         Text(
                             modifier = modifier.align(Alignment.CenterHorizontally),
-                            text = "신청 승인하시겠습니까?",
+                            text = "신청 거부하시겠습니까?",
                             color = colors.BLACK,
                             style = type.bodyLarge,
                             fontSize = 18.sp
@@ -96,6 +94,7 @@ fun BitgoeulApproveDialog(
                                     style = type.bodySmall
                                 )
                             }
+
                             Box(
                                 modifier = modifier
                                     .clip(
@@ -107,9 +106,9 @@ fun BitgoeulApproveDialog(
                                         )
                                     )
                                     .background(
-                                        color = colors.P5
+                                        color = colors.E5
                                     )
-                                    .weight(1f),
+                                    .weight(1f)
                             ) {
                                 Text(
                                     modifier = modifier
@@ -118,9 +117,9 @@ fun BitgoeulApproveDialog(
                                         .clickable {
                                             onQuit()
                                         },
-                                    text = stringResource(id = R.string.approve),
+                                    text = stringResource(id = R.string.reject),
                                     color = colors.WHITE,
-                                    style = type.bodySmall,
+                                    style = type.bodySmall
                                 )
                             }
                         }
@@ -129,14 +128,4 @@ fun BitgoeulApproveDialog(
             }
         }
     }
-}
-
-@Preview
-@Composable
-fun BitgoeulApproveDialogPre() {
-    BitgoeulApproveDialog(
-        content = "국가는 국민 모두의 생산 및 생활의 기반이 되는 국토의 효율적이고 균형있는 이용·개발과 보전을 위하여 법률이 정하는 바에 의하여 그에 관한 필요 김하온 박주홍 강민수 이동욱 정찬교 이정우 김새미 서주미 정윤서 ",
-        isVisible = true,
-        onQuit = {}
-    )
 }
